requirements.txt
Pillow==9.4.0
praw==7.7.0
pyimgur==0.6.0
stability_sdk==0.3.2
